Fix signed/unsigned comparisons
- Comparisons between size_t/ptrdiff_t require special handling.  The
size_t is strictly larger than ptrdiff_t, so if ptrdiff_t > 0, then we
can safely convert it to size_t.

+++ SurgSim/Physics/UnitTests/Fem1DMechanicalValidationTests.cpp
@@ -104,7 +104,7 @@ public:
 
 	void addBoundaryCondition(size_t node, size_t dof)
 	{
-		for (int i = 0; i < dof; i++)
+		for (size_t i = 0; i < dof; i++)
 		{
 			boundaryConditions.push_back(std::make_pair(node, i));
 		}
