+++ SurgSim/Physics/UnitTests/Fem2DElementTriangleTests.cpp
@@ -1093,7 +1093,7 @@ TEST_F(Fem2DElementTriangleTests, ForceAndMatricesAPITest)
 	forceVector.setZero();
 	ones.setOnes();
 	tri->addMatVec(m_restState, 1.0, 0.0, 0.0, ones, &forceVector);
-	for (int rowId = 0; rowId < numDof; rowId++)
+	for (size_t rowId = 0; rowId < numDof; rowId++)
 	{
 		SCOPED_TRACE("Test addMatVec API with Mass component only");
 		EXPECT_NEAR(expectedMassMatrix.row(rowId).sum(), forceVector[rowId], epsilon);
@@ -1101,7 +1101,7 @@ TEST_F(Fem2DElementTriangleTests, ForceAndMatricesAPITest)
 	// Test addMatVec API with Damping component only
 	forceVector.setZero();
 	tri->addMatVec(m_restState, 0.0, 1.0, 0.0, ones, &forceVector);
-	for (int rowId = 0; rowId < numDof; rowId++)
+	for (size_t rowId = 0; rowId < numDof; rowId++)
 	{
 		SCOPED_TRACE("Test addMatVec API with Damping component only");
 		EXPECT_NEAR(0.0, forceVector[rowId], epsilon);
@@ -1109,7 +1109,7 @@ TEST_F(Fem2DElementTriangleTests, ForceAndMatricesAPITest)
 	// Test addMatVec API with Stiffness component only
 	forceVector.setZero();
 	tri->addMatVec(m_restState, 0.0, 0.0, 1.0, ones, &forceVector);
-	for (int rowId = 0; rowId < numDof; rowId++)
+	for (size_t rowId = 0; rowId < numDof; rowId++)
 	{
 		SCOPED_TRACE("Test addMatVec API with Stiffness component only");
 		EXPECT_NEAR(expectedStiffnessMatrix.row(rowId).sum(), forceVector[rowId], epsilon);
@@ -1117,7 +1117,7 @@ TEST_F(Fem2DElementTriangleTests, ForceAndMatricesAPITest)
 	// Test addMatVec API with mix Mass/Damping/Stiffness components
 	forceVector.setZero();
 	tri->addMatVec(m_restState, 1.0, 2.0, 3.0, ones, &forceVector);
-	for (int rowId = 0; rowId < numDof; rowId++)
+	for (size_t rowId = 0; rowId < numDof; rowId++)
 	{
 		SCOPED_TRACE("Test addMatVec API with mix Mass/Damping/Stiffness components");
 		double expectedCoef = 1.0 * expectedMassMatrix.row(rowId).sum() +
